Adicionando aula 03 - Variáveis globais e objects
