--- pacotes-e-composicao/src/main/kotlin/br/com/thais/bytebank/modelo/ContaCorrente.kt
@@ -1,16 +1,0 @@
-package br.com.thais.bytebank.modelo
-
-class ContaCorrente(
-    titular: Client,
-    numero: Int
-) : ContaTransferivel(
-    titular = titular,
-    numero = numero
-) {
-    override fun saca(valor: Double) {
-        if(this.saldo >= valor) {
-            this.saldo -= valor
-        }
-    }
-
-}
