--- pacotes-e-composicao/src/main/kotlin/br/com/thais/bytebank/modelo/ContaPoupanca.kt
@@ -1,17 +1,0 @@
-package br.com.thais.bytebank.modelo
-
-class ContaPoupanca(
-    titular: Client,
-    numero: Int
-) : ContaTransferivel(
-    titular = titular,
-    numero = numero
-) {
-
-    override fun saca(valor: Double) {
-        val valorComTaxa = valor + 0.1
-        if(this.saldo >= valorComTaxa) {
-            this.saldo -= valorComTaxa
-        }
-    }
-}
